+++ ChannelIO/Source/Models/CHColors.swift
@@ -7,41 +7,41 @@
 //
 
 import Foundation
-import UIColor_Hex_Swift
+import HexColors
 
 struct CHColors {
-  static let dark = UIColor("#516378")
-  static let deepDark = UIColor("#3F4F5F")
+  static let dark = UIColor("#516378")!
+  static let deepDark = UIColor("#3F4F5F")!
   static let stale10 = UIColor(red: 81/255,
                                green: 99/255, blue: 120/255, alpha: 0.1)
-  static let warmPink = UIColor("#ff3d69")
-  static let gray = UIColor("#7C858C")
+  static let warmPink = UIColor("#ff3d69")!
+  static let gray = UIColor("#7C858C")!
   static let white = UIColor.white
   static let black = UIColor.black
-  static let yellowishOrange = UIColor("#ffa908")
-  static let lightGray = UIColor("#eef1f4")
+  static let yellowishOrange = UIColor("#ffa908")!
+  static let lightGray = UIColor("#eef1f4")!
   static let darkTwo = UIColor(red: 81/255,
                                green: 99/255, blue: 120/255, alpha: 0.2)
-  static let shamrockGreen = UIColor("#00d63d")
-  static let blueyGrey = UIColor("#98a7b3")
+  static let shamrockGreen = UIColor("#00d63d")!
+  static let blueyGrey = UIColor("#98a7b3")!
   static let blueyGrey60 = UIColor(red: 152/255,
                                    green: 167/255, blue: 179/255, alpha: 0.6)
-  static let charcoalGrey = UIColor("#2b2f35")
-  static let minicGray = UIColor("#1a516377")
-  static let cobalt = UIColor("#00A6FF")
-  static let light = UIColor("#8696A4")
-  static let azure = UIColor("#00a6ff")
-  static let lightAzure = UIColor("#e6f9ff")
-  static let snow = UIColor("#F3F4F5")
-  static let lightSnow = UIColor("#fbfbfc")
-  static let yellow = UIColor("#ffa908")
-  static let defaultTint = UIColor("#007aff")
-  static let dark10 = UIColor("#3f152128")
-  static let paleGrey = UIColor("#eaeff4")
-  static let iceBlue = UIColor("#ecf9ff")
-  static let paleSkyBlue = UIColor("#b7e7ff")
-  static let white15 = UIColor("#ffffff26")
-  static let white40 = UIColor("#ffffff66")
-  static let dark5 = UIColor("#1521280c")
-  static let dark20 = UIColor("#15212833")
+  static let charcoalGrey = UIColor("#2b2f35")!
+  static let minicGray = UIColor("#1a516377")!
+  static let cobalt = UIColor("#00A6FF")!
+  static let light = UIColor("#8696A4")!
+  static let azure = UIColor("#00a6ff")!
+  static let lightAzure = UIColor("#e6f9ff")!
+  static let snow = UIColor("#F3F4F5")!
+  static let lightSnow = UIColor("#fbfbfc")!
+  static let yellow = UIColor("#ffa908")!
+  static let defaultTint = UIColor("#007aff")!
+  static let dark10 = UIColor("#3f152128")!
+  static let paleGrey = UIColor("#eaeff4")!
+  static let iceBlue = UIColor("#ecf9ff")!
+  static let paleSkyBlue = UIColor("#b7e7ff")!
+  static let white15 = UIColor("#ffffff26")!
+  static let white40 = UIColor("#ffffff66")!
+  static let dark5 = UIColor("#1521280c")!
+  static let dark20 = UIColor("#15212833")!
 }